#321 Deleted config.genomes entirely, moved url2db, db2url to accession.py.
